+++ src/DeviceLayout.jl
@@ -365,6 +365,7 @@ include("polygons.jl")
 import .Polygons:
     Polygon,
     Ellipse,
+    LineSegment,
     Circle,
     ClippedPolygon,
     RelativeRounded,
@@ -390,6 +391,7 @@ import .Polygons:
 export Polygons,
     Polygon,
     Ellipse,
+    LineSegment,
     Circle,
     ClippedPolygon,
     RelativeRounded,