+++ src/polygons.jl
@@ -63,7 +63,7 @@ import IntervalTrees: IntervalTree, IntervalValue
 import IntervalSets.(..)
 import IntervalSets.endpoints
 
-export Polygon, ClippedPolygon, Ellipse, Circle
+export Polygon, ClippedPolygon, Ellipse, Circle, LineSegment
 export circle,
     circle_polygon,
     clip,
@@ -1517,11 +1517,15 @@ end
 
 ### cutting algorithm
 
-abstract type D1{T} end
+abstract type D1{T} <: GeometryEntity{T} end
 Δy(d1::D1) = d1.p1.y - d1.p0.y
 Δx(d1::D1) = d1.p1.x - d1.p0.x
 
 ab(p0, p1) = Point(gety(p1) - gety(p0), getx(p0) - getx(p1))
+
+to_polygons(::D1{T}) where {T} = Polygon{T}[]
+
+transform(d1::T, f::Transformation) where {T <: D1} = T(f(d1.p0), f(d1.p1))
 
 """
     LineSegment{T} <: D1{T}