+++ src/schematics/schematics.jl
@@ -1151,6 +1151,14 @@ end
 
 function index_layer!(sch::Schematic, ly, nextindex=1)
     index_components = ComponentNode[]
+    # Index top-level schematic elements
+    for (idx, m) in pairs(element_metadata(sch.coordinate_system))
+        if (layer(m) == ly)
+            element_metadata(sch.coordinate_system)[idx] = SemanticMeta(m, index=nextindex)
+            nextindex = nextindex + 1
+        end
+    end
+    # Index component-level elements
     for node in nodes(sch.graph)
         startindex = nextindex
         nextindex = index_layer!(coordsys(sch[node]), component(node), ly, nextindex)